% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeimmoniums.r
\name{analyze_immoniums}
\alias{analyze_immoniums}
\title{Extracting isotope peak intensities for immonium ions from mzML file}
\usage{
analyze_immoniums(file, width = 0.001, ions = immoniumIons, fixSigma = T)
}
\arguments{
\item{file}{mzML file containing immonium scans (m/z range 50-200, HCD energy 50);}

\item{width}{fitting window width;}

\item{ions}{list of compounds to fit. Each element of the list should represent
named vector with element names as vector names and numbers of corresponding
elements as value, e.g. \code{c("C" = 4, "H" = 8, "N" = 1)};}

\item{fixSigma}{if True, \code{sigma} is fitted only for monoisotopic peak
and fixed for other peaks;}
}
\value{
Data frame with fitting results for each spectrum in input \code{file}.
}
\description{
Extracting isotope peak intensities for immonium ions from mzML file
}
