% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_isopeaks.R
\name{fit_isopeaks}
\alias{fit_isopeaks}
\title{fit_isopeaks}
\usage{
fit_isopeaks(x, y, startPars = c(rc = 0.02, rn = 0.003, rh = 4e-04, dm = 0),
  I = 10000, sigma = 0.001, mu = 70, maxAttempts = 1000)
}
\arguments{
\item{x}{vector of x-values;}

\item{y}{vector of y-values;}

\item{startPars}{names vector of initial parameters
\code{mu}, \code{I} and \code{sigma};}

\item{sigma}{if \code{sigma>0} provided, two-parameter fitting is performed,
fixing \code{sigma} at provided value;}

\item{maxAttempts}{maximum number of attempts, see below;}
}
\value{
Returns an object of class \code{nls} if the fitting was successful.
}
\description{
Fits gaussian peak for the data
\code{ y = I/sigma * exp(-1/2*(x-mu)^2 / sigma^2) }
using nls function with initial parameters specified by named vector
\code{startPars}. If \code{sigma > 0} orivuded, then only two parameters are
fitted using nls - intensity \code{I} and position \code{mu}.
}
\examples{
fitPeak(x=x, y=y, startPars=c('mu' = 70.065, 'I' = 5e3, 'sigma' = 3e-4))
fitPeak(x=x, y=y, startPars=c('mu' = 70.065, 'I' = 5e3, 'sigma' = 3e-4), sigma=3e-4)
}
