% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIsoPeaks.R
\name{get_isopeaks2}
\alias{get_isopeaks2}
\title{Isotopic peaks fitting for specified m/z}
\usage{
get_isopeaks2(ss, mz, width = 0.002, npoint = 10, tol = 0.01,
  fixSigma = T, plot = F)
}
\arguments{
\item{ss}{Spectrum formatted as numeric matrix with 2 columns (m/z and i),
sorted by mz}

\item{mz}{m/z value corresponding to the monoisotopic mass of singly charged
ion}

\item{width}{m/z interval for searching the isotopic peaks. Default value is
0.002 which is good for resolution 50-60K (at m/z=200).}

\item{npoint}{Minimum number of data points in the peak. If peak have less
points it will be rejected.}

\item{tol}{m/z tolerance for searching monoisotopic peak.}

\item{fixSigma}{If True, then peak width is only fitted for monoisotopic
peak, and then it is fixed for the rest isotopes.}

\item{plot}{If True, fitted peaks are plotted using ggplot.}
}
\value{
Dataframe with peak parameters
}
\description{
Petforms peak fitting at provided 'm/z' value, and also fits fine isotopic
structure of +1 isotopic peak corresponding to the 13C, 15N and 2H isotopes.
Unlike \code{get_isopeaks}, isotopic peaks are not fitted separately, but
rather as a single envelope with fixed m/z values for peaks.
Returns data_frame object containing information about fitting results.
}
