% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monoMass.R
\name{monoMass}
\alias{monoMass}
\title{monoMass function calculates the monoisotopic mass of a molecule with a
composition \code{comp}}
\usage{
monoMass(comp)
}
\arguments{
\item{comp}{- named vector of molecule composition, e.g.
\code{c('C' = 4, 'H' = 8, 'N' = 1)}}
}
\value{
monoisotopic mass of the molecule
}
\description{
monoMass function calculates the monoisotopic mass of a molecule with a
composition \code{comp}
}
\examples{
monoMass(c('C' = 4, 'H' = 8, 'N' = 1)) # 70.06567
}
