% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSquared.R
\name{rSquared}
\alias{rSquared}
\title{rSquared}
\usage{
rSquared(model, y)
}
\arguments{
\item{model}{object of class nls or lm, or similar providing \code{predict}
function.}

\item{y}{numeric vector of dependent variable used to make \code{model}.}
}
\value{
single numeric R^2 value, or NA if \code{class(model) == "try-error"}.
}
\description{
calculates R^2 for the fitting model
}
