% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeImmoniums.R
\name{summarizeImmoniums}
\alias{summarizeImmoniums}
\title{summarizeImmoniums - generate the summary of the immonium analysis
Given the results of \code{analyzeImmoniums} function, produces a nice html
report}
\usage{
summarizeImmoniums(data = NA, files = NA, group = ifelse(is.na(data),
  "file", NA), resultPath = "./isoMS_result")
}
\arguments{
\item{data}{- a dataframe returned by \code{analyzeImmoniums} function.
If \code{NA} then using \code{files}}

\item{files}{- vector of csv-files containing results of \code{analyzeImmoniums}}

\item{group}{- name of the column in data}
}
\description{
summarizeImmoniums - generate the summary of the immonium analysis
Given the results of \code{analyzeImmoniums} function, produces a nice html
report
}
